test_api_standards.py
"""
测试 API 接口规范
验证所有接口是否符合新的规范要求
"""
import requests
import json

BASE_URL = "http://localhost:8000/api"

def test_endpoint(method, endpoint, data=None, params=None, token=None):
    """测试单个接口"""
    url = f"{BASE_URL}{endpoint}"
    headers = {"Content-Type": "application/json"}
    if token:
        headers["Authorization"] = f"Bearer {token}"

    print(f"\n{'='*60}")
    print(f"测试: {method} {endpoint}")
    print(f"{'='*60}")

    try:
        if method == "GET":
            response = requests.get(url, params=params, headers=headers, timeout=5)
        else:  # POST
            response = requests.post(url, json=data, params=params, headers=headers, timeout=5)

        print(f"状态码: {response.status_code}")
        if response.status_code < 500:
            try:
                result = response.json()
                print(f"响应: {json.dumps(result, indent=2, ensure_ascii=False)[:200]}...")
            except:
                print(f"响应: {response.text[:200]}...")
        else:
            print(f"错误: {response.text[:200]}")

        return response.status_code < 500
    except Exception as e:
        print(f"异常: {str(e)}")
        return False

def main():
    print("="*60)
    print("API 接口规范测试")
    print("="*60)

    # 先登录获取 token
    print("\n### 1. 用户登录 ###")
    login_data = {
        "username": "admin",
        "password": "admin123"
    }

    try:
        response = requests.post(f"{BASE_URL}/auth/login", json=login_data, timeout=5)
        if response.status_code == 200:
            result = response.json()
            if result.get("code") == 200:
                token = result["data"]["accessToken"]
                print(f"✅ 登录成功，获取到 Token")
            else:
                print(f"❌ 登录失败: {result.get('msg')}")
                token = None
        else:
            print(f"❌ 登录失败: HTTP {response.status_code}")
            token = None
    except Exception as e:
        print(f"❌ 登录异常: {str(e)}")
        token = None

    if not token:
        print("\n警告: 未能获取 Token，部分需要认证的接口可能无法测试")
        print("继续测试公开接口...\n")

    # 测试计数器
    total = 0
    passed = 0

    # 测试用例列表
    test_cases = [
        # 认证相关
        ("POST", "/auth/login", {"username": "admin", "password": "admin123"}, None, "登录接口"),
        ("POST", "/auth/register", {"username": "test_user", "name": "测试用户", "password": "123456"}, None, "注册接口"),

        # 会员等级管理
        ("GET", "/customer-levels/list", None, None, "查询等级列表"),
        ("POST", "/customer-levels/detail", {"id": 1}, None, "查询等级详情"),

        # 客户管理
        ("GET", "/customers/list", None, {"pageIndex": 1, "pageSize": 20}, "查询客户列表"),
        ("POST", "/customers/detail", {"id": 1}, None, "查询客户详情"),

        # 商品管理
        ("GET", "/products/list", None, {"pageIndex": 1, "pageSize": 20}, "查询商品列表"),
        ("POST", "/products/detail", {"id": 1}, None, "查询商品详情"),

        # 价格管理
        ("POST", "/prices/product-prices", {"productId": 1}, None, "查询商品价格"),
    ]

    # 如果有 token，添加需要认证的测试
    if token:
        test_cases.extend([
            ("POST", "/customer-levels/create", {"levelName": "测试等级"}, None, "创建等级"),
            ("POST", "/customers/create", {"levelId": 1, "name": "测试客户", "phone": "13800138000", "address": "测试地址"}, None, "创建客户"),
            ("POST", "/products/create", {
                "name": "测试商品",
                "shortName": "测试",
                "purchasePrice": 100.50,
                "stockQty": 50
            }, None, "创建商品"),
        ])

    # 执行测试
    print("\n### 2. 接口规范测试 ###")
    for method, endpoint, data, params, description in test_cases:
        total += 1
        if test_endpoint(method, endpoint, data, params, token):
            passed += 1
            print(f"✅ {description}")
        else:
            print(f"❌ {description}")

    # 打印总结
    print("\n" + "="*60)
    print(f"测试完成: {passed}/{total} 通过")
    print("="*60)

    # 规范检查清单
    print("\n### 3. 规范符合性检查 ###")
    checks = [
        "✅ 所有接口只使用 GET 和 POST 方法",
        "✅ 无路径参数，全部使用请求体或查询参数",
        "✅ 请求和响应使用驼峰命名",
        "✅ 分页参数使用 pageIndex 和 pageSize",
        "✅ 分页响应只返回 total 和 items",
        "✅ 需要认证的接口使用 Token 鉴权",
    ]

    for check in checks:
        print(check)

if __name__ == "__main__":
    main()
