pyproject.toml
[project]
name = "shop-backend"
version = "0.1.0"
description = "超市后端管理系统"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.17",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
